requirements.txt
boto3==1.36.1
dj-rest-auth==7.0.1
Django==5.0
django-allauth==65.3.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
python-dotenv